expr: avoid int truncation warning for 'loopcount' assignment
MSVC says:

  C:\GitLab-Runner\builds\graphviz\graphviz\lib\expr\exeval.c(1072,28): warning
    C4244: '=': conversion from '__int64' to 'int', possible loss of data
    [C:\GitLab-Runner\builds\graphviz\graphviz\lib\expr\expr.vcxproj]

In practice, it is infeasible to create a GVPR loop with > INT_MAX loop
operations. But aligning these types is a simple way to quiet the compiler.

+++ lib/expr/exlib.h
@@ -115,7 +115,7 @@ typedef struct Print_s			/* compiled printf arg node	*/
 	int		errors;		/* fatal error count		*/ \
 	int		formals;	/* parsing formal args		*/ \
 	int		linewrap;	/* linep wrapped around line[]	*/ \
-	int		loopcount;	/* break|continue|return count	*/ \
+	long long loopcount; /* break|continue|return count	*/ \
 	int		loopop;		/* break|continue|return op	*/ \
 	int		nesting;	/* exstatement() nesting	*/
